python/requirements.txt
clickhouse-driver==0.2.8
clickhouse-cityhash==1.0.2.4
clickhouse-connect==0.7.18
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.2
boto3==1.34.101
huggingface-hub==0.24.5
transformers==4.43.1
sentence-transformers==2.7.0
tqdm==4.66.4
rich==13.7.1
